start moving host functions to new component
