+++ pokeR/ClientApp/src/app/shared-components/host-controls/host-controls.component.html
@@ -1,0 +1,26 @@
+<div class="container">
+  host controls here! :)
+  <div *ngIf="player.isHost || currentTagline" class="mb-3 d-flex flex-row justify-content-center">
+    <span *ngIf="!player.isHost" class="h5">{{currentTagline}}</span>
+    <div class="input-group" *ngIf="player.isHost">
+      <input type="text" class="form-control bg-light" (input)="onTextChange($event.target.value)"
+        [value]="currentTagline">
+      <input type="number" class="form-control col-2 bg-light text-right" placeholder="Countdown" min="0" max="600"
+        [(ngModel)]="countdownInput" [disabled]="countdownIsActive">
+      <div class="input-group-append">
+        <button class="btn btn-outline-secondary" type="button" placement="bottom" ngbTooltip="Start Countdown"
+          [disabled]="countdownInput==null || countdownIsActive" (click)="startCountdown()">
+          <i class="mdi mdi-timer"></i>
+        </button>
+        <button class="btn" type="button" placement="bottom" ngbTooltip="Start New Round" [disabled]="countdownIsActive"
+          (click)="startNewRound()">
+          <i class="mdi mdi-restart"></i>
+        </button>
+      </div>
+    </div>
+    <div *ngIf="countdownIsActive && getCountdownValue()" class="mt-3">
+      <ngb-progressbar [value]="getCountdownValue()">
+        {{getTimeText()}}
+      </ngb-progressbar>
+    </div>
+  </div>
