+++ pokeR/ClientApp/src/app/shared-components/roster/roster.component.html
@@ -1,56 +1,10 @@
-<!--<div class="bg-light text-muted py-3">
-  <div class="container">
-    <div *ngIf="player.isHost || currentTagline" class="mb-3 d-flex flex-row justify-content-center">
-      <span *ngIf="!player.isHost" class="h5">{{currentTagline}}</span>
-      <div class="input-group" *ngIf="player.isHost">
-        <input type="text" class="form-control bg-light" (input)="onTextChange($event.target.value)"
-          [value]="currentTagline">
-        <input type="number" class="form-control col-2 bg-light text-right" placeholder="Countdown" min="0" max="600"
-          [(ngModel)]="countdownInput" [disabled]="countdownIsActive">
-        <div class="input-group-append">
-          <button class="btn btn-outline-secondary" type="button" placement="bottom" ngbTooltip="Start Countdown"
-            [disabled]="countdownInput==null || countdownIsActive" (click)="startCountdown()">
-            <i class="mdi mdi-timer"></i>
-          </button>
-          <button class="btn" [class.btn-outline-secondary]="!readyToStart" [class.btn-primary]="readyToStart"
-            type="button" placement="bottom" ngbTooltip="Start New Round" [disabled]="countdownIsActive"
-            (click)="startNewRound()">
-            <i class="mdi mdi-restart"></i>
-          </button>
-        </div>
-      </div>
-    </div>
-    <div class="d-flex flex-row justify-content-around flex-wrap">
-      <div *ngFor="let u of users" class="d-flex flex-column align-items-center user-animate"
-        (dragenter)="onDragEnter(u)" (dragleave)="onDragLeave()" (drop)="setKing(u.id)"
-        (dragover)="allowDrop($event, u)" [class.text-warning]="u==hoveredUser" [class.drop-hover]="u==hoveredUser">
-        <div class="h5 mb-0">
-          <i *ngIf="!u.currentCardId" class="mdi mdi-checkbox-blank-circle-outline"
-            [class.text-danger]="getRemainingUsers() == 1"></i>
-          <i *ngIf="u.currentCardId" class="mdi mdi-checkbox-marked-circle mdi text-success"></i>
-        </div>
-        <div>
-          <i *ngIf="u.isHost" [draggable]="player.isHost" [class.draggable]="player.isHost"
-            [class.forbidden]="!player.isHost" class="mdi mdi-crown" (dragend)="onDragLeave()"></i>
-          <span [class.font-weight-bold]="player.id == u.id">{{u.displayName}}</span>
-        </div>
-      </div>
-    </div>
-    <div *ngIf="countdownIsActive && getCountdownValue()" class="mt-3">
-      <ngb-progressbar [value]="getCountdownValue()">
-        {{getTimeText()}}
-      </ngb-progressbar>
-    </div>
-  </div>
-</div>-->
 <app-pane-drawer label="Roster" direction="left">
   <div class="player-list">
     <div *ngFor="let u of users" class="player user-animate" (dragenter)="onDragEnter(u)" (dragleave)="onDragLeave()"
       (drop)="setKing(u.id)" (dragover)="allowDrop($event, u)" [class.drop-hover]="u==hoveredUser"
       [class.played]="u.currentCardId">
       <div class="icon">
-        <i *ngIf="!u.currentCardId" class="mdi mdi-checkbox-blank-circle-outline"
-          [class.text-danger]="getRemainingUsers() == 1"></i>
+        <i *ngIf="!u.currentCardId" class="mdi mdi-checkbox-blank-circle-outline"></i>
         <i *ngIf="u.currentCardId" class="mdi mdi-checkbox-marked-circle"></i>
       </div>
       <div>
